requirements.txt
Pillow==9.2.0
python-docx==0.8.11
pywinauto==0.6.6

